backend/requirements.txt
asgiref==3.8.1
boto3==1.34.141
botocore==1.34.141
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.6
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
fonttools==4.52.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
s3transfer==0.10.2
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
wcwidth==0.2.13
Werkzeug==3.0.3
